perf: stabilize SemanticCache benchmark and migrate DDG to ddgs
- Increase SemanticCache benchmark warm-up iterations to 5 and samples to 20
- Use P95 latency threshold (100ms) and relaxed max (200ms) for benchmarks
- Migrate DuckDuckGo search from duckduckgo-search to ddgs package (>= 1.0.0)
- Update dependencies and mocks across root and skill directories
- Fix mobile/tablet E2E tests by handling sidebar visibility and backdrop interception
- Update CI quality gate to include E2E test results
