+++ .github/workflows/ci-ui.yml
@@ -140,9 +140,8 @@ jobs:
   web-e2e:
     name: E2E Tests (web)
     runs-on: ubuntu-latest
-    timeout-minutes: 45
+    timeout-minutes: 30
     needs: [web-build]
-    if: github.actor != 'dependabot[bot]'
     steps:
       - uses: actions/checkout@v6
 
@@ -171,7 +170,7 @@ jobs:
         run: |
           npm run start &
           npx wait-on http://localhost:3000 -t 60000
-          npx playwright test --project=desktop --project=mobile --project=tablet --workers=3
+          npx playwright test --project=desktop --project=mobile --project=tablet
         working-directory: web
         env:
           BASE_URL: http://localhost:3000
