+++ README.md
@@ -438,25 +438,13 @@ cd web && npx playwright test --ui
 │   └── ui/                # Design system (tokens, components, Storybook)
 ├── web/                   # Next.js web UI (Vercel)
 │   ├── app/
-│   │   ├── page.tsx       # Resolver form (356 lines)
-│   │   ├── components/    # Extracted components
-│   │   │   ├── Sidebar.tsx
-│   │   │   ├── MainContent.tsx
-│   │   │   ├── KeyboardShortcutsModal.tsx
-│   │   │   ├── History.tsx
-│   │   │   ├── ResultCard.tsx
-│   │   │   └── ProfileCombobox.tsx
-│   │   ├── constants.ts   # Profile/provider constants
-│   │   ├── middleware.ts  # Rate limiting
-│   │   ├── error.tsx      # Error boundary
-│   │   ├── not-found.tsx  # 404 page
+│   │   ├── page.tsx       # Resolver form
 │   │   └── help/page.tsx  # Help & FAQ
-│   ├── lib/               # Utilities (circuit-breaker, errors, quality, keys, routing)
-│   ├── tests/             # Unit tests (Vitest) + E2E (Playwright)
+│   ├── tests/e2e/         # Playwright E2E tests
 │   └── vercel.json        # Deployment config
 ├── tests/                 # Python test suite
 ├── docs/                  # Standards, examples
-├── plans/                 # Roadmap & audit (see plans/AUDIT.md)
+├── plans/                 # Roadmap & audit
 ├── .agents/skills/        # Canonical skill definitions
 │   └── do-web-doc-resolver/
 │       ├── SKILL.md       # Main skill file
