+++ pyproject.toml
@@ -39,7 +39,7 @@ all = [
     "tavily-python>=0.3.0",
     "firecrawl-py>=0.0.5",
     "mistralai>=1.0.0",
-    "ddgs>=2.0.0",
+    "ddgs>=1.0.0",
     "diskcache>=5.6.0",
 ]
