+++ web/tests/e2e/app.spec.ts
@@ -37,6 +37,16 @@ async function mockAppState(page: import("@playwright/test").Page): Promise<void
   });
 }
 
+// Helper to open sidebar on mobile viewports
+async function openSidebarIfMobile(page: import("@playwright/test").Page): Promise<void> {
+  const openMenuButton = page.getByRole("button", { name: "Open menu" });
+  if (await openMenuButton.isVisible()) {
+    await openMenuButton.click();
+    // Wait for transition animation
+    await page.waitForTimeout(500);
+  }
+}
+
 // Helper to wait for app to be fully loaded (async state initialization)
 async function waitForApp(page: import("@playwright/test").Page): Promise<void> {
   await mockAppState(page);
@@ -530,6 +540,7 @@ test.describe("Navigation", () => {
 test.describe("Collapsible Sidebar", () => {
   test("sidebar is visible by default", async ({ page }) => {
     await waitForApp(page);
+    await openSidebarIfMobile(page);
     await expect(page.getByTestId("sidebar-toggle")).toBeVisible();
     await expect(page.locator("text=Configuration")).toBeVisible();
     await expect(page.locator("label").filter({ hasText: "Profile" })).toBeVisible();
@@ -537,6 +548,7 @@ test.describe("Collapsible Sidebar", () => {
 
   test("sidebar collapses when clicking Configuration header", async ({ page }) => {
     await waitForApp(page);
+    await openSidebarIfMobile(page);
     await expect(page.locator("label").filter({ hasText: "Profile" })).toBeVisible();
     await page.getByTestId("sidebar-toggle").click();
     await expect(page.locator("label").filter({ hasText: "Profile" })).not.toBeVisible();
@@ -544,6 +556,7 @@ test.describe("Collapsible Sidebar", () => {
 
   test("sidebar expands when clicking Configuration header again", async ({ page }) => {
     await waitForApp(page);
+    await openSidebarIfMobile(page);
     await page.getByTestId("sidebar-toggle").click();
     await expect(page.locator("label").filter({ hasText: "Profile" })).not.toBeVisible();
     await page.getByTestId("sidebar-toggle").click();
@@ -552,6 +565,7 @@ test.describe("Collapsible Sidebar", () => {
 
   test("toggle label shows correct text", async ({ page }) => {
     await waitForApp(page);
+    await openSidebarIfMobile(page);
     await expect(page.getByTestId("sidebar-toggle").locator("text=Hide")).toBeVisible();
     await page.getByTestId("sidebar-toggle").click();
     await expect(page.getByTestId("sidebar-toggle").locator("text=Show")).toBeVisible();
@@ -559,6 +573,7 @@ test.describe("Collapsible Sidebar", () => {
 
   test("Keys link is visible in sidebar header", async ({ page }) => {
     await waitForApp(page);
+    await openSidebarIfMobile(page);
     const keysLink = page.locator('a[href="/settings"]');
     await expect(keysLink).toBeVisible();
     await expect(keysLink).toContainText("Keys");
@@ -566,6 +581,7 @@ test.describe("Collapsible Sidebar", () => {
 
   test("Keys link navigates to settings", async ({ page }) => {
     await waitForApp(page);
+    await openSidebarIfMobile(page);
     await page.locator('a[href="/settings"]').click();
     await expect(page).toHaveURL(/\/settings/);
   });
@@ -574,12 +590,14 @@ test.describe("Collapsible Sidebar", () => {
 test.describe("Collapsible API Keys", () => {
   test("API Keys section is collapsed by default", async ({ page }) => {
     await waitForApp(page);
+    await openSidebarIfMobile(page);
     await expect(page.getByTestId("api-keys-toggle")).toBeVisible();
     await expect(page.locator("label").filter({ hasText: "Serper" })).not.toBeVisible();
   });
 
   test("API Keys section expands on click", async ({ page }) => {
     await waitForApp(page);
+    await openSidebarIfMobile(page);
     await page.getByTestId("api-keys-toggle").click();
     await expect(page.locator("label").filter({ hasText: "Serper" })).toBeVisible();
     await expect(page.locator("label").filter({ hasText: "Tavily" })).toBeVisible();
@@ -587,6 +605,7 @@ test.describe("Collapsible API Keys", () => {
 
   test("API Keys section collapses on second click", async ({ page }) => {
     await waitForApp(page);
+    await openSidebarIfMobile(page);
     await page.getByTestId("api-keys-toggle").click();
     await expect(page.locator("label").filter({ hasText: "Serper" })).toBeVisible();
     await page.getByTestId("api-keys-toggle").click();
@@ -597,6 +616,7 @@ test.describe("Collapsible API Keys", () => {
 test.describe("Profile Provider Indicators", () => {
   test("profile providers are shown as active by default", async ({ page }) => {
     await waitForApp(page);
+    await openSidebarIfMobile(page);
     // Free profile is default: exa_mcp (DuckDuckGo may be disabled if Mistral is active)
     // Exa MCP should have the active style (green border)
     const exaButton = page.locator("button").filter({ hasText: "Exa MCP" });
@@ -608,12 +628,14 @@ test.describe("Profile Provider Indicators", () => {
 
   test("profile status text shows provider count", async ({ page }) => {
     await waitForApp(page);
+    await openSidebarIfMobile(page);
     // Free profile shows exa_mcp, may or may not show DuckDuckGo depending on Mistral gating
     await expect(page.locator("text=Using free profile")).toBeVisible();
   });
 
   test("clicking a provider switches to custom selection", async ({ page }) => {
     await waitForApp(page);
+    await openSidebarIfMobile(page);
     // Click Exa MCP (free, always available) to toggle it off the profile
     await page.locator("button").filter({ hasText: "Exa MCP" }).click();
     // Status should change to custom selection
@@ -622,6 +644,7 @@ test.describe("Profile Provider Indicators", () => {
 
   test("manual selection overrides profile display", async ({ page }) => {
     await waitForApp(page);
+    await openSidebarIfMobile(page);
     // Click Exa MCP (free, always available) to manually toggle
     const exaButton = page.locator("button").filter({ hasText: "Exa MCP" });
     await exaButton.click();
