--- web/tests/quality.test.ts
@@ -1,125 +1,0 @@
-import { describe, expect, it } from "vitest";
-import { scoreContent } from "../lib/quality";
-
-describe("scoreContent", () => {
-  describe("tooShort", () => {
-    it("marks short content as too short", () => {
-      const result = scoreContent("short");
-      expect(result.tooShort).toBe(true);
-    });
-
-    it("marks long content as not too short", () => {
-      const result = scoreContent("a".repeat(600));
-      expect(result.tooShort).toBe(false);
-    });
-
-    it("penalizes score for short content", () => {
-      const short = scoreContent("short");
-      const long = scoreContent("a".repeat(600));
-      expect(short.score).toBeLessThan(long.score);
-    });
-  });
-
-  describe("missingLinks", () => {
-    it("marks content with no links as missing links", () => {
-      const result = scoreContent("long content without links");
-      expect(result.missingLinks).toBe(true);
-    });
-
-    it("marks content with links as having links", () => {
-      const result = scoreContent("content", ["https://example.com"]);
-      expect(result.missingLinks).toBe(false);
-    });
-
-    it("penalizes score for missing links", () => {
-      const noLinks = scoreContent("a".repeat(600));
-      const withLinks = scoreContent("a".repeat(600), ["https://example.com"]);
-      expect(noLinks.score).toBeLessThan(withLinks.score);
-    });
-  });
-
-  describe("duplicateHeavy", () => {
-    it("detects duplicate content", () => {
-      // Need enough lines to trigger duplicateHeavy check (>5 lines)
-      const duplicateContent = "same line\nsame line\nsame line\nsame line\nsame line\nsame line\n";
-      const result = scoreContent(duplicateContent);
-      expect(result.duplicateHeavy).toBe(true);
-    });
-
-    it("does not mark unique content as duplicate heavy", () => {
-      const uniqueContent = "line one\nline two\nline three\nline four\nline five\nline six\n";
-      const result = scoreContent(uniqueContent);
-      expect(result.duplicateHeavy).toBe(false);
-    });
-
-    it("penalizes score for duplicates", () => {
-      const duplicate = scoreContent("same\nsame\nsame\nsame\nsame\nsame\n");
-      const unique = scoreContent("one\ntwo\nthree\nfour\nfive\nsix\n");
-      expect(duplicate.score).toBeLessThan(unique.score);
-    });
-  });
-
-  describe("noisy", () => {
-    it("does not mark content as noisy (max 5 signals, threshold > 6)", () => {
-      // Implementation has 5 signals but threshold is > 6, so noisy is never true
-      const noisyContent = "cookie subscribe javascript log in sign up";
-      const result = scoreContent(noisyContent);
-      expect(result.noisy).toBe(false); // Can never exceed threshold with only 5 signals
-    });
-
-    it("does not mark clean content as noisy", () => {
-      const cleanContent = "Documentation about React hooks and state management.";
-      const result = scoreContent(cleanContent);
-      expect(result.noisy).toBe(false);
-    });
-
-    it("counts all 5 signals without triggering noisy", () => {
-      const allSignals = "cookie subscribe javascript log in sign up";
-      const result = scoreContent(allSignals);
-      // noiseCount = 5, threshold > 6, so noisy = false
-      expect(result.noisy).toBe(false);
-    });
-  });
-
-  describe("acceptable", () => {
-    it("marks good content as acceptable", () => {
-      const result = scoreContent("a".repeat(600), ["https://example.com"]);
-      expect(result.acceptable).toBe(true);
-    });
-
-    it("marks short content as not acceptable even with good score", () => {
-      const result = scoreContent("short", ["https://example.com"]);
-      expect(result.acceptable).toBe(false);
-    });
-
-    it("marks content below threshold as not acceptable", () => {
-      const result = scoreContent("a".repeat(600)); // missing links penalty
-      expect(result.acceptable).toBe(false);
-    });
-  });
-
-  describe("score range", () => {
-    it("returns score between 0 and 1", () => {
-      const result = scoreContent("any content");
-      expect(result.score).toBeGreaterThanOrEqual(0);
-      expect(result.score).toBeLessThanOrEqual(1);
-    });
-
-    it("returns high score for ideal content without duplicates", () => {
-      // Need >500 chars, many unique lines, and links to get high score
-      const uniqueLines = Array.from({ length: 12 }, (_, i) => `This is line number ${i + 1} with unique content for testing.`).join("\n");
-      const result = scoreContent(uniqueLines, ["https://a.com", "https://b.com"]);
-      // 12 lines with 12 unique should avoid duplicateHeavy (< max(5, 6) = 6)
-      // > 500 chars should avoid tooShort
-      expect(result.tooShort).toBe(false);
-      expect(result.duplicateHeavy).toBe(false);
-      expect(result.score).toBeGreaterThanOrEqual(0.65);
-    });
-
-    it("returns low score for worst content", () => {
-      // Short content gets multiple penalties
-      const result = scoreContent("short");
-      expect(result.score).toBeLessThanOrEqual(0.25);
-    });
-  });
-});
